requirements.txt
# Core
python-dotenv>=1.0.0

# LangChain stack (1.x)
langchain>=1.0,<2.0
langchain-core>=1.0,<2.0
langchain-classic>=1.0,<2.0
langchain-community>=0.4,<1.0
langchain-text-splitters>=1.0,<2.0
langchain-chroma>=1.0,<2.0
langchain-huggingface>=1.0,<2.0
langchain-anthropic>=1.0,<2.0
langchain-ollama>=1.0,<2.0

# Vector store + embeddings
chromadb>=0.5
sentence-transformers>=3.0

# Document processing
pymupdf4llm>=0.0.17
pymupdf>=1.24
ebooklib>=0.18
beautifulsoup4>=4.12
lxml>=5.0
python-docx>=1.0
rank_bm25>=0.2.2

# UI
chainlit>=2.0,<3.0

# Utilities
tqdm>=4.66